Soma reais resolvido

+++ capitulo02/ex01/soma_reais.cpp
@@ -1,4 +1,4 @@
-//Criar um programa em linguagem C que efetue
+//01 - Criar um programa em linguagem C que efetue
 //a soma de dois números dos tipo real.
 
 #include <windows.h>
@@ -7,15 +7,15 @@
 main(){
 	system("cls");
 	
-	int num1, num2, soma;
+	float num1, num2, soma;
 	
 	printf("Programa de soma\n");
 	printf("Digite os dois numeros para soma:\n");
-	scanf("%d%d",&num1,&num2);
+	scanf("%f%f",&num1,&num2);
 	
 	soma = num1 + num2;
 	
-	printf("Soma = %d\n",soma);
+	printf("Soma = %f\n",soma);
 	
 	system("pause");
 }
